requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
geopy==2.3.0
Pillow==9.4.0
psycopg2-binary==2.9.7
requests==2.31.0
SQLAlchemy==1.4.46
Werkzeug==2.2.5
